build-jedi-cache.sh: managed to make it build actual cache
    Did not test whether the completion actually works though.

+++ build-jedi-cache.sh
@@ -1,5 +1,49 @@
 #!/bin/bash
 
+PYTHON_EXECUTABLE=python
+
+_help() {
+    cat <<EOF
+Usage: $0 [options...]
+
+Options:
+    -3      Use python3 instead of python as a Python interpreter
+    -p INTERPRETER
+            Specify custom Python interpreter path
+    -h      Show this help message
+EOF
+}
+
+while getopts '3p:h' arg; do
+    case "$arg" in
+        3) PYTHON_EXECUTABLE=python3;;
+        p) PYTHON_EXECUTABLE="$OPTARG";;
+        h) _help; exit 0;;
+        *) _help >&2; exit 127;;
+    esac
+done
+
+shift $(( $OPTIND - 1 ))
+
+# if $PYTHON_EXECUTABLE does not contain slashes,
+# then use /usr/bin/env in shebang line
+# otherwise, try getting the absolute path and use it instead
+if [ "$(expr "$PYTHON_EXECUTABLE" : '.*/')" -eq 0 ]; then
+    # also, find the real 'env'
+    _env=
+    for _p in /bin/env /usr/bin/env; do
+        if [ -x "$_p" ]; then
+            _env=$_p
+            break
+        fi
+    done
+    [ "$_env" ] || { echo "Could not find 'env' executable" >&2; exit 4; }
+
+    SHEBANG="#!$_env $PYTHON_EXECUTABLE"
+else
+    SHEBANG="#!$(realpath "$PYTHON_EXECUTABLE")"
+fi
+
 VIM_PLUGINS_DIR="$HOME/.vim/bundle"
 if [ ! -d $VIM_PLUGINS_DIR ]; then
     VIM_PLUGINS_DIR="$HOME/.vim/plugged"
@@ -9,13 +53,25 @@ if [ ! -d $VIM_PLUGINS_DIR ]; then
     exit 2
 fi
 
-JEDI_PATH=$VIM_PLUGINS_DIR/YouCompleteMe/third_party/ycmd/third_party/jedi
+JEDI_PATH=
+
+for _p in \
+    $VIM_PLUGINS_DIR/YouCompleteMe/third_party/ycmd/third_party/jedi \
+    $VIM_PLUGINS_DIR/YouCompleteMe/third_party/ycmd/third_party/JediHTTP/vendor/jedi
+do
+    if [ -d "$_p" ]; then
+        JEDI_PATH=$_p
+        break
+    fi
+done
+
+[ "$JEDI_PATH" ] || { echo "Unable to find Jedi directory"; exit 3; }
 
 #python fakegir.py
 rm -rf $HOME/.cache/jedi
 
 pkgs=""
-echo '#/bin/env python' >/tmp/fakeprg.py
+echo "$SHEBANG" >/tmp/fakeprg.py
 for f in ~/.cache/fakegir/gi/repository/*.py; do
     pkgname=`basename $f | cut -d . -f 1`
     pkgs="$pkgs $pkgname"
@@ -33,5 +89,6 @@ for pkgname in $pkgs; do
 
     export PYTHONPATH=$HOME/.cache/fakegir
     pushd $JEDI_PATH
-    $JEDI_PATH/sith.py -f run completions /tmp/${pkgname}-fakeprg.py $fakeprg_lines $compl_col
+    $PYTHON_EXECUTABLE $JEDI_PATH/sith.py \
+        -f run completions /tmp/${pkgname}-fakeprg.py $fakeprg_lines $compl_col
 done
